Render Linux dynamic UI textures
Add CPU-backed lock/unlock storage for Linux/OpenGL Texture2D objects and upload supported formats through the native GL path. Feed decoded DDS, generated bitmap, cursor, and font atlas data through that storage so real UI image and text quads bind native textures.

Tighten Linux texture validity around supported CPU-backed resources and keep packed colors 32-bit on 64-bit Linux to avoid row-size corruption.

Windows/full-render paths remain on their existing Direct3D material and texture implementations.

+++ pw/branches/r1117/Src/System/Color.h
@@ -15,7 +15,7 @@ namespace Render
 		union 
 		{
 			struct{ TValue B, G, R, A; }; // intel byte order aka little-endian 
-			unsigned long Dummy;
+			unsigned int Dummy; // Keep packed color 32-bit on 64-bit Linux.
 		};
 		
     #ifdef WIN32
